perf(audio): reduce peak memory per song from ~1.5 GB to ~300 MB
H1 (audio.py): _stft_mag — replace giant int64 idx array with
  as_strided view — eliminates ~212 MB indexing array per call,
  7+ calls per song. Zero copy.

H2 (separate.py): remove 2 unnecessary .astype(np.float32) copies
  — saves 12 MB + 106 MB per chunk (model_in and output).

H2b (audio.py): vocals.mean(dtype=np.float32) avoids float64
  intermediate.

M1 (audio.py): load_mono_mix — sum incrementally instead of
  holding all stems simultaneously. Peak: 312 MB → 52 MB for 6 stems.

M2 (audio.py): load_mono .mogg path — stream 8-byte header first,
  don't read entire file into bytes.

M3 (audio.py): _try_mdx_separation — del mono/stereo when no longer
  needed so GC doesn't see them as live references.

Tests: 163 passed

+++ downcharter/separate.py
@@ -342,7 +342,7 @@ def separate_vocals(
         model_in = np.stack(
             [S_real[0], S_imag[0], S_real[1], S_imag[1]], axis=0
         )  # (4, dim_f, dim_t)
-        model_in = model_in[np.newaxis, :, :, :].astype(np.float32)  # (1, 4, dim_f, dim_t)
+        model_in = model_in[np.newaxis, :, :, :]  # (1, 4, dim_f, dim_t)  — already float32
 
         # ── ONNX inference ────────────────────────────────────────────
         try:
@@ -388,4 +388,4 @@ def separate_vocals(
     # Clip to float32 range
     np.clip(out, -1.0, 1.0, out=out)
 
-    return out.astype(np.float32)
+    return out  # already float32
